CRUD 코드이그나이터로 이식완료

+++ application/views/article/article_detail.php
@@ -12,13 +12,21 @@
 </div>
 
 <button>
-    <a href="/article/articlecontroller/articleEdit/<?php echo $article->id; ?>">
+    <a href="/article/articlecontroller/articleEdit/<?php echo $article->id; ?>" onclick="return confirmEdit();">
     <?php echo $article->id; ?>번 게시글 수정하기
     </a>
 </button>
 
 <button>
-    <a href="/article/articlecontroller/processArticleDelete/<?php echo $article->id; ?>">
+    <a href="/article/articlecontroller/processArticleDelete/<?php echo $article->id; ?>" onclick="return confirmDelete();">
     <?php echo $article->id; ?>번 게시글 삭제하기
     </a>
 </button>
+<script>
+    function confirmEdit() {
+        return confirm('정말로 수정하시겠습니까?');
+    }
+    function confirmDelete() {
+        return confirm('삭제된 게시글은 다시 되돌릴 수 없습니다.\n정말로 삭제하시겠습니까?');
+    }
+</script>
